pyproject.toml
[tool.poetry]
name = "image-translator"
version = "0.1.0"
description = ""
authors = ["Giulia Caglia <cagliagiulia42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kaggle = "^1.6.12"
click = "^8.1.7"
torchvision = "^0.18.0"
torch = "^2.3.0"
pillow = "^10.3.0"
numpy = "^1.26.4"
dill = "^0.3.8"
matplotlib = "^3.9.0"
mlflow = "^2.13.0"


[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
isort = "^5.13.2"
black = "^24.4.2"
pre-commit = "^3.7.0"
ipython = "^8.24.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
fetch-data = "image_translator.scripts.fetch_data:fetch_data"
train = "image_translator.scripts.train:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
